pyproject.toml
[project]
name = "radar-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "loguru>=0.7.3",
    "pyqtgraph>=0.13.7",
    "pyside6-essentials>=6.8.2.1",
    "tomli>=2.2.1 ; python_full_version < '3.11'",
]

[dependency-groups]
dev = [
    "matplotlib>=3.10.0",
    "scikit-learn>=1.6.0",
    "seaborn>=0.13.2",
]
